Refatoração completa e Modais
Refatoração do sistema, componentizando toda a classe ficha, separando todas as funcionalidades em classes diferentes.

Criação de modais para confirmação de ações, substituindo alertas do navegador
