+++ js/components/actions/acoes.html
@@ -1,0 +1,60 @@
+<div id="view-combate" class="tab-content">
+    <div style="display: flex; justify-content: space-between; align-items: center;">
+        <h1>Ações de Combate</h1>
+        <button id="btnNovaAcao" style="background: var(--color-power); color: white; border: none; padding: 10px 20px; border-radius: 8px; font-weight: bold; cursor: pointer;">+ Nova Ação</button>
+    </div>
+    <p style="color: #999; font-size: 0.9rem; margin-bottom: 20px;">Gerencie suas técnicas e ataques aqui.</p>
+
+    <div class="section-header">
+        <i class="fas fa-gavel" style="color: var(--color-padrao);"></i>
+        <h3>Ações Padrão</h3>
+    </div>
+    <div id="lista-padrao" class="action-grid"></div>
+
+    <div class="section-header">
+        <i class="fas fa-bolt" style="color: var(--color-bonus);"></i>
+        <h3>Ações Bônus</h3>
+    </div>
+    <div id="lista-bonus" class="action-grid"></div>
+
+    <div class="section-header">
+        <i class="fas fa-bomb" style="color: var(--color-power);"></i>
+        <h3>Ações Poderosas</h3>
+    </div>
+    <div id="lista-power" class="action-grid"></div>
+    
+    <div class="section-header">
+        <i class="fas fa-shield-alt" style="color: var(--color-react);"></i>
+        <h3>Reações</h3>
+    </div>
+    <div id="lista-react" class="action-grid"></div>
+</div>
+
+<div id="modalAcao" class="modal-overlay" style="display: none;">
+    <div class="modal-box"> 
+        <h2>Nova Ação</h2>
+        <label>Nome da Ação</label>
+        <input type="text" id="newActionName" placeholder="Ex: Soco Meteoro">
+        <label>Descrição</label>
+        <textarea id="newActionDesc" placeholder="Descreva o efeito..." style="width: 100%; margin-top: 5px;"></textarea>
+        <div style="display: grid; grid-template-columns: 1fr 1fr; gap: 10px; margin-top: 10px;">
+            <div>
+                <label>Tipo</label>
+                <select id="newActionType" style="width: 100%; padding: 8px;">
+                    <option value="padrao">Ação Padrão</option>
+                    <option value="bonus">Ação Bônus</option>
+                    <option value="power">Ação Poderosa</option>
+                    <option value="react">Reação</option>
+                </select>
+            </div>
+            <div>
+                <label>Dano / Tag</label>
+                <input type="text" id="newActionTag" placeholder="Ex: 1d8+2">
+            </div>
+        </div>
+        <div style="margin-top: 20px; text-align: right;">
+            <button id="btnFecharModal" style="background: #ccc; color: black; border: none; padding: 10px 20px; border-radius: 8px; font-weight: bold; cursor: pointer;">Cancelar</button>
+            <button id="btnSalvarAcao" style="margin-left: 10px; background: var(--color-padrao); color: white; border: none; padding: 10px 20px; border-radius: 8px; font-weight: bold; cursor: pointer;">Salvar</button>
+        </div>
+    </div>
+</div>
