+++ js/components/profile/perfil.html
@@ -1,0 +1,89 @@
+<aside class="sidebar" style="position: relative;">
+    <div style="position: absolute; top: 20px; right: 20px; display: flex; gap: 10px;">
+        <i id="btnToggleTheme" class="fas fa-moon" 
+           style="color: #aaa; cursor: pointer; font-size: 1.2rem; transition: 0.3s;">
+        </i>
+        <i id="btnEditarFicha" class="fas fa-cog" 
+           style="color: #aaa; cursor: pointer; font-size: 1.2rem; transition: 0.3s;">
+        </i>
+    </div>
+    <style> 
+        #btnEditarFicha:hover { color: var(--primary) !important; transform: rotate(90deg); } 
+        #btnToggleTheme:hover { color: var(--color-bonus) !important; transform: scale(1.1); }
+    </style>
+
+    <div class="profile-header">
+        <div class="avatar-circle">
+            <img src="https://api.dicebear.com/7.x/adventurer/svg?seed=Felix" alt="Avatar">
+        </div>
+        <h2 class="char-name" id="displayNome">Carregando...</h2>
+        
+        <span class="char-level">
+            Nível <span id="valNivel" class="editable-num">1</span>
+        </span>
+    </div>
+
+    <hr style="border: 0; border-top: 1px solid #eee;">
+
+    <div class="bar-group bar-hp">
+        <div class="bar-label">
+            <span>❤️ Vida</span>
+            <div class="stat-values">
+                <span id="valHp" class="editable-num">0</span> / <span id="maxHp">0</span>
+            </div>
+        </div>
+        <div class="progress-track"><div class="progress-fill" id="fillHp" style="width: 50%;"></div></div>
+    </div>
+
+    <div class="bar-group bar-en">
+        <div class="bar-label">
+            <span>⚡ Pontos de Poder</span>
+            <div class="stat-values">
+                <span id="valPp" class="editable-num">0</span> / <span id="maxPp">0</span>
+            </div>
+        </div>
+        <div class="progress-track"><div class="progress-fill" id="fillPp" style="width: 80%;"></div></div>
+    </div>
+
+    <div class="bar-label" style="margin-top: 15px; color: var(--color-power);">ATRIBUTOS</div>
+    <div class="attr-grid">
+        <div class="attr-card"><div class="attr-label">FOR</div><div class="attr-value" id="attr-forca">0</div></div>
+        <div class="attr-card"><div class="attr-label">DES</div><div class="attr-value" id="attr-destreza">0</div></div>
+        <div class="attr-card"><div class="attr-label">CON</div><div class="attr-value" id="attr-constituicao">0</div></div>
+        <div class="attr-card"><div class="attr-label">INT</div><div class="attr-value" id="attr-sabedoria">0</div></div> 
+        <div class="attr-card"><div class="attr-label">VON</div><div class="attr-value" id="attr-vontade">0</div></div>
+        <div class="attr-card"><div class="attr-label">PRE</div><div class="attr-value" id="attr-presenca">0</div></div>
+    </div>
+    
+    <button id="btnSair" style="margin-top: auto; padding: 10px; border:none; background: #eee; cursor: pointer; border-radius: 6px;">Sair</button>
+</aside>
+
+<div id="modalFicha" class="modal-overlay" style="display: none;">
+    <div class="modal-box">
+        <h2>Editar Atributos</h2>
+        <p style="font-size: 0.8rem; color: #666; margin-bottom: 15px;">Ajuste seus limites e atributos base.</p>
+        <div style="display: flex; gap: 10px;">
+            <div style="flex: 1;">
+                <label>Vida Máxima</label>
+                <input type="number" id="editHpMax">
+            </div>
+            <div style="flex: 1;">
+                <label>PP Máximo</label>
+                <input type="number" id="editPpMax">
+            </div>
+        </div>
+        <label style="margin-top: 15px; border-bottom: 1px solid #eee; padding-bottom: 5px;">Atributos</label>
+        <div class="grid-atributos" style="display: grid; grid-template-columns: 1fr 1fr 1fr; gap: 10px; margin-top: 10px;">
+            <div><label style="font-size: 0.7rem;">FOR</label><input type="number" id="editFor"></div>
+            <div><label style="font-size: 0.7rem;">DES</label><input type="number" id="editDes"></div>
+            <div><label style="font-size: 0.7rem;">CON</label><input type="number" id="editCon"></div>
+            <div><label style="font-size: 0.7rem;">INT</label><input type="number" id="editSab"></div>
+            <div><label style="font-size: 0.7rem;">VON</label><input type="number" id="editVon"></div>
+            <div><label style="font-size: 0.7rem;">PRE</label><input type="number" id="editPre"></div>
+        </div>
+        <div style="margin-top: 20px; text-align: right;">
+            <button id="btnFecharFicha" style="background: #ccc; color: black; border: none; padding: 10px 20px; border-radius: 8px; font-weight: bold; cursor: pointer;">Cancelar</button>
+            <button id="btnSalvarFicha" style="margin-left: 10px; background: var(--primary); color: white; border: none; padding: 10px 20px; border-radius: 8px; font-weight: bold; cursor: pointer;">Salvar</button>
+        </div>
+    </div>
+</div>
