+++ js/utils/firebase-config.js
@@ -1,0 +1,18 @@
+import { initializeApp } from "https://www.gstatic.com/firebasejs/10.7.1/firebase-app.js";
+import { getDatabase } from "https://www.gstatic.com/firebasejs/10.7.1/firebase-database.js";
+import { getAuth } from "https://www.gstatic.com/firebasejs/10.7.1/firebase-auth.js";
+
+const firebaseConfig = {
+    apiKey: "AIzaSyBkp8ZUYMCfRokbpMl2fBGTvfMxzzvgaeY",
+    authDomain: "rpgo-onepiece.firebaseapp.com",
+    databaseURL: "https://rpgo-onepiece-default-rtdb.firebaseio.com",
+    projectId: "rpgo-onepiece",
+    storageBucket: "rpgo-onepiece.firebasestorage.app",
+    messagingSenderId: "726770644982",
+    appId: "1:726770644982:web:7c06f46940cc5142c3f9d7",
+    measurementId: "G-HSBMTMB5XK"
+};
+
+const app = initializeApp(firebaseConfig);
+export const db = getDatabase(app);
+export const auth = getAuth(app);
